+++ lib/src/main/java/com/smileidentity/compose/consent/bvn/OrchestratedBvnConsentScreen.kt
@@ -1,9 +1,11 @@
 package com.smileidentity.compose.consent.bvn
 
+import android.content.pm.ActivityInfo
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
+import com.smileidentity.compose.components.ForceOrientation
 import com.smileidentity.viewmodel.BvnConsentScreens
 import com.smileidentity.viewmodel.BvnConsentViewModel
 import com.smileidentity.viewmodel.viewModelFactory
@@ -18,6 +20,7 @@ internal fun OrchestratedBvnConsentScreen(
     ),
     successfulBvnVerification: () -> Unit,
 ) {
+    ForceOrientation(ActivityInfo.SCREEN_ORIENTATION_PORTRAIT)
     val uiState by viewModel.uiState.collectAsStateWithLifecycle()
     when (uiState.bvnConsentScreens) {
         BvnConsentScreens.BvnInputScreen -> BvnInputScreen(userId = userId)