+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Bump Sentry to 6.29.0
 
 ### Removed
+- `filename` property from `PrepUploadRequest`, as it is no longer required
 
 ## 10.0.0-beta07